requirements.txt
python-pdfbox
beautifulsoup4
fake-useragent
selenium
pypdf
huggingface_hub
acryl-datahub[datahub-rest]
datahub
torch
transformers
python-dotenv
langchain
tiktoken
git+https://github.com/SergeyPirogov/webdriver_manager@master #webdriver-manager
mlflow==2.12.2
greenplum-python
torchvision
accelerate

matplotlib
pmdarima
prophet
psycopg2-binary
sqlalchemy
python-dotenv
pandas
wheel
scikit-learn
nbconvert
seaborn
traitlets
scipy
plotly
watchdog
mpld3
streamlit==1.18.0
Pillow
streamlit-autorefresh
jupyterflow
lz4
pyyaml==5.3.1
prometheus-client
protobuf # protobuf==3.20.0 # protobuf==3.19.5
fastapi
uvicorn[standard]
gunicorn
boto3
tensorflow
rsocket
python-snappy
filelock==3.9.0
pyarrow<7.0.0
hickle
python-multipart
evidently
pysqlite3-binary
kfp==2.5.0
python-multipart
pyservicebinding
download_and_extract
altair==4.0
sentencepiece
urllib3<2
GitPython
pydantic==2.6.0
tf-keras
onnxruntime
onnx
ort-nightly
--extra-index-url=https://aiinfra.pkgs.visualstudio.com/PublicPackages/_packaging/ORT-Nightly/pypi/simple/
--pre
onnxruntime-genai
--extra-index-url=https://aiinfra.pkgs.visualstudio.com/PublicPackages/_packaging/onnxruntime-genai/pypi/simple/
--pre
.